[REMOVE] unnecessary  files
