+++ HANE24/View/Calendar/CalendarBodyView.swift
@@ -20,6 +20,7 @@ struct CalendarBodyView: View {
     }
 }
 
+
 #Preview {
     CalendarBodyView(selectedDate: .constant(Date()))
 }
