+++ HANE24/View/Calendar/CalendarGridView.swift
@@ -115,6 +115,21 @@ struct CalendarGridView: View {
         }
     }
     
+    func calculateLogColor(accumulationTime: Int64) -> Color {
+        switch accumulationTime {
+        case 0:
+            return .backgroundCalendar
+        case 1 ... 14400:
+            return .dateBackgroundLevel1
+        case 14400 ... 28800:
+            return .dateBackgroundLevel2
+        case 28800 ... 43200:
+            return .dateBackgroundLevel3
+        default:
+            return .dateBackgroundLevel4
+        }
+    }
+
 }
